Move testApplabConsoleOutput helper to levelTestHelpers
